src/utils/multilingual.py
import torch
import torch.nn as nn
from typing import Dict, List, Optional, Tuple
import logging
from collections import defaultdict
import fasttext
import numpy as np
from transformers import XLMRobertaTokenizer
from sacremoses import MosesTokenizer, MosesDetokenizer

class LanguageManager:
    """
    Manages language-specific operations and resources
    """
    def __init__(self, config):
        self.config = config
        self.logger = logging.getLogger(__name__)
        
        # Language detection model
        self.lang_detector = fasttext.load_model('lid.176.bin')
        
        # Initialize tokenizers
        self.xlm_tokenizer = XLMRobertaTokenizer.from_pretrained(config.model_name)
        self.moses_tokenizers = {
            lang: MosesTokenizer(lang=lang)
            for lang in config.languages
        }
        self.moses_detokenizers = {
            lang: MosesDetokenizer(lang=lang)
            for lang in config.languages
        }
        
        # Language-specific resources
        self.initialize_resources()
        
    def initialize_resources(self):
        """Initialize language-specific resources"""
        self.stop_words = defaultdict(set)
        self.special_tokens = defaultdict(set)
        self.language_codes = {
            'en': 'english',
            'fr': 'french',
            'es': 'spanish',
            'it': 'italian',
            'de': 'german'
        }
        
        # Load language resources
        for lang in self.config.languages:
            self._load_language_resources(lang)
            
    def _load_language_resources(self, language: str):
        """Load resources for specific language"""
        try:
            # Load stop words
            from nltk.corpus import stopwords
            self.stop_words[language] = set(stopwords.words(self.language_codes[language]))
            
            # Load special tokens
            self._load_special_tokens(language)
            
        except Exception as e:
            self.logger.warning(f"Could not load all resources for {language}: {e}")
            
    def _load_special_tokens(self, language: str):
        """Load special tokens for language"""
        # Common special tokens
        special_tokens = {
            'en': {'[URL]', '[EMAIL]', '[MENTION]', '[HASHTAG]'},
            'fr': {'[URL]', '[EMAIL]', '[MENTION]', '[HASHTAG]'},
            'es': {'[URL]', '[EMAIL]', '[MENTION]', '[HASHTAG]'},
            'it': {'[URL]', '[EMAIL]', '[MENTION]', '[HASHTAG]'},
            'de': {'[URL]', '[EMAIL]', '[MENTION]', '[HASHTAG]'}
        }
        
        self.special_tokens[language] = special_tokens.get(language, set())

class MultilingualCalibration:
    """
    Handles cross-lingual calibration and alignment
    """
    def __init__(self, config):
        self.config = config
        
        # Language-specific scaling
        self.language_scaling = nn.ParameterDict({
            lang: nn.Parameter(torch.ones(config.hidden_size))
            for lang in config.languages
        })
        
        # Cross-lingual alignment
        self.alignment_layer = nn.Linear(config.hidden_size, config.hidden_size)
        self.layer_norm = nn.LayerNorm(config.hidden_size)
        
        # Language embeddings
        self.language_embeddings = nn.Embedding(
            len(config.languages),
            config.hidden_size
        )
        
        self.dropout = nn.Dropout(config.dropout)
        
    def forward(
        self,
        hidden_states: torch.Tensor,
        language: str,
        attention_mask: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Apply language-specific calibration
        
        Args:
            hidden_states: Input tensor [batch_size, seq_len, hidden_size]
            language: Language code
            attention_mask: Optional attention mask
        Returns:
            Calibrated tensor
        """
        # Apply language-specific scaling
        if language in self.language_scaling:
            scale = self.language_scaling[language]
            hidden_states = hidden_states * scale.unsqueeze(0).unsqueeze(0)
            
        # Add language embeddings
        lang_idx = self.config.languages.index(language)
        lang_embedding = self.language_embeddings(
            torch.tensor(lang_idx, device=hidden_states.device)
        )
        hidden_states = hidden_states + lang_embedding.unsqueeze(0).unsqueeze(0)
        
        # Apply cross-lingual alignment
        aligned_states = self.alignment_layer(hidden_states)
        calibrated_states = self.layer_norm(aligned_states)
        
        return self.dropout(calibrated_states)

class LanguageDetector:
    """
    Handles language detection and validation
    """
    def __init__(self, config):
        self.config = config
        self.lang_detector = fasttext.load_model('lid.176.bin')
        self.min_confidence = config.language_detection_threshold
        
    def detect_language(
        self,
        text: str
    ) -> Tuple[str, float]:
        """
        Detect language of text
        
        Args:
            text: Input text
        Returns:
            Tuple of (language code, confidence)
        """
        predictions = self.lang_detector.predict(text, k=1)
        lang_code = predictions[0][0].replace('__label__', '')
        confidence = predictions[1][0]
        
        return lang_code, confidence
        
    def validate_language(
        self,
        text: str,
        expected_language: str
    ) -> bool:
        """
        Validate if text is in expected language
        
        Args:
            text: Input text
            expected_language: Expected language code
        Returns:
            True if validation passes
        """
        detected_lang, confidence = self.detect_language(text)
        
        # Check confidence threshold
        if confidence < self.min_confidence:
            return False
            
        # Check language match
        return detected_lang == expected_language

class MultilingualTokenizer:
    """
    Handles multilingual tokenization
    """
    def __init__(self, config):
        self.config = config
        
        # XLM-RoBERTa tokenizer
        self.xlm_tokenizer = XLMRobertaTokenizer.from_pretrained(
            config.model_name
        )
        
        # Language-specific tokenizers
        self.moses_tokenizers = {
            lang: MosesTokenizer(lang=lang)
            for lang in config.languages
        }
        
    def tokenize(
        self,
        text: str,
        language: str,
        add_special_tokens: bool = True,
        max_length: Optional[int] = None
    ) -> Dict:
        """
        Tokenize text for specific language
        
        Args:
            text: Input text
            language: Language code
            add_special_tokens: Whether to add special tokens
            max_length: Optional maximum length
        Returns:
            Dictionary of tokenization outputs
        """
        # Normalize text for language
        if language in self.moses_tokenizers:
            text = self.moses_tokenizers[language].tokenize(text)
            text = ' '.join(text)
            
        # Tokenize with XLM-RoBERTa
        outputs = self.xlm_tokenizer(
            text,
            add_special_tokens=add_special_tokens,
            max_length=max_length,
            truncation=True,
            return_attention_mask=True,
            padding='max_length' if max_length else False
        )
        
        return outputs

class MultilingualAugmenter:
    """
    Handles multilingual data augmentation
    """
    def __init__(self, config):
        self.config = config
        self.language_manager = LanguageManager(config)
        
    def augment_text(
        self,
        text: str,
        language: str,
        techniques: Optional[List[str]] = None
    ) -> List[str]:
        """
        Apply augmentation techniques
        
        Args:
            text: Input text
            language: Language code
            techniques: List of augmentation techniques to apply
        Returns:
            List of augmented texts
        """
        if techniques is None:
            techniques = ['substitute', 'insert', 'delete']
            
        augmented_texts = []
        
        for technique in techniques:
            if technique == 'substitute':
                augmented = self._substitute_words(text, language)
            elif technique == 'insert':
                augmented = self._insert_words(text, language)
            elif technique == 'delete':
                augmented = self._delete_words(text, language)
            else:
                continue
                
            augmented_texts.extend(augmented)
            
        return augmented_texts
        
    def _substitute_words(
        self,
        text: str,
        language: str
    ) -> List[str]:
        """Substitute words with synonyms"""
        # Implement word substitution logic
        pass
        
    def _insert_words(
        self,
        text: str,
        language: str
    ) -> List[str]:
        """Insert relevant words"""
        # Implement word insertion logic
        pass
        
    def _delete_words(
        self,
        text: str,
        language: str
    ) -> List[str]:
        """Delete words while maintaining meaning"""
        # Implement word deletion logic
        pass