requirements.txt
# Core dependencies
torch>=1.9.0
transformers>=4.12.0
datasets>=1.11.0
numpy>=1.19.5
pandas>=1.3.0
tqdm>=4.62.0

# NLP tools
sentencepiece>=0.1.96
tokenizers>=0.10.3
sacremoses>=0.0.45
nltk>=3.6.3
spacy>=3.1.0
gensim>=4.1.2
fasttext-wheel>=0.9.2
stop-words>=2018.7.23
language-detector>=5.0.2
symspellpy>=6.7.0
seqeval>=1.2.2

# Text processing
PyYAML>=5.4.1
sacrebleu>=2.0.0
subword-nmt>=0.3.7
sentencepiece>=0.1.96
ftfy>=6.0.3

# Machine learning
scikit-learn>=0.24.2
scipy>=1.7.1
higher>=0.2.1
torchcrf>=1.1.0

# Metrics and logging
wandb>=0.12.0
tensorboard>=2.7.0
matplotlib>=3.4.3
seaborn>=0.11.2

# Development tools
pytest>=6.2.5
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910

# Utilities
pyyaml>=5.4.1
jsonschema>=3.2.0
requests>=2.26.0
typing-extensions>=3.10.0
dataclasses>=0.8; python_version < '3.7'

# Optional dependencies for full functionality
# Uncomment if needed
# apex>=0.1.0  # For mixed precision training
# torch-scatter>=2.0.8  # For graph operations
# pytorch-lightning>=1.4.9  # For structured training loops

# Language specific models
# After installing spacy, run:
# python -m spacy download en_core_web_sm
# python -m spacy download fr_core_news_sm
# python -m spacy download es_core_news_sm
# python -m spacy download it_core_news_sm
# python -m spacy download de_core_news_sm